+++ src/plugins/lists/base/AddFab/AddEntryDialog.tsx
