+++ src/plugins/lists/base/EntryList/InjectEntryDialog.spec.tsx
