+++ src/plugins/lists/base/EntryList/InjectEntryDialog.tsx
