+++ src/plugins/lists/base/EntryList/RemoveEntryDialog.tsx
