+++ src/plugins/lists/base/RemoveListDialog.spec.tsx
