+++ src/plugins/lists/base/RemoveListDialog.tsx
