+++ src/plugins/lists/base/hooks/list.ts
@@ -1,7 +1,7 @@
 import { useReducer, Reducer, useEffect, useCallback, useMemo } from 'react';
 import { action } from 'utils/hooks/actions';
 import { createContainer, useContainer } from 'unstated-next';
-import { usePluginContainer } from 'plugins/managedList/hooks/api';
+import { usePluginContainer } from './api';
 import { List, AddListRequest } from '../types';
 
 export const enum Constants {
