+++ src/plugins/lists/base/index.tsx
@@ -1,10 +1,10 @@
 import React, { FC } from 'react';
 import { Theme } from '@material-ui/core';
 import { css } from '@emotion/core';
+import { useInjectPageTitle } from 'core/layout/AppBar/hooks';
 import { NoPaddingWrapper } from 'common/styles';
 import { ListContainer } from './hooks/list';
 import Entries from './Entries';
-import { Entry } from './types';
 
 export const content = (theme: Theme) => css`
   flex: 1;
@@ -16,14 +16,22 @@ export const content = (theme: Theme) => css`
   }
 `;
 
-const ManagedList: FC = <T extends Entry>() => (
-  <ListContainer.Provider>
-    <NoPaddingWrapper>
-      <div css={content}>
-        <Entries<T> />
-      </div>
-    </NoPaddingWrapper>
-  </ListContainer.Provider>
-);
+interface Props {
+  title: string;
+}
+
+const ManagedList: FC<Props> = ({ title }) => {
+  useInjectPageTitle(title);
+
+  return (
+    <ListContainer.Provider>
+      <NoPaddingWrapper>
+        <div css={content}>
+          <Entries />
+        </div>
+      </NoPaddingWrapper>
+    </ListContainer.Provider>
+  );
+};
 
 export default ManagedList;
